added support for primvar setup in slum, by assign the input=True to a parameter fixed slum node refresh when parameters got connections. Now attributes are refreshed and maintain the connections.
added some more sanityCheck. Now slumMaya plugin checks all unknown nodes in a scene if they are slum nodes.
If so, it figures the path for the workspace project were the slum template for the unknown node is defined, and
can automatically switch to that path, refresh the plugin and reload the scene.