COMPATIBILITY WARNING:
starting on this version, all parameters defined in the parameters method are automatically converted by the 3delight module
to 3delight input parameters. NO NEED TO DUPLICATE THE PARAMETERS CREATION ON THE delight METHOD ANYMORE. the delight renderer method
still returns a tupple for compatibility reasons only, but I'm evaluating the need for that. I'm aiming on completely remove 
the tupple return and make the renderer method return ONLY shader code. This way ALL parameters will be specified in the parameters method!

there's a new "primvar" attribute to a parameter wich flags it as a primtive variable input. If combined with output attribute, it becomes
a output primitive variable (a secondary pass in 3delight, for example)

all example shaders are still using the old DUPLICATED version of parameters. I'll fix it asap.

bugfixes:

fixed a lot of bugs related to stability in maya.
now a node can be refreshed even if there are connections to attributes. all the connections will be kept.