added bakery.slum template. A shader to bake maps in 3delight. This bakery template is quite interesting. The delight method can figure if it's running in maya. If so, it figures WHAT is connected to the input color parameters, and bake only those. It also figures the name of the connected node and assign it as the suffix of the baked file.
fixed string parameters which was causing the export of the shader to fail.
added <project> and <scene> keyword substitution to slumNode. (need to transform this into a generic slum method which is called 
automatically by the parameter retrieve method of the slum shaderbase class. Clients can override the function as needed.)

started to add a "file" ui type.
fixed a looooot of little bugs. 
TODO: fix maya template refresh when a float attr is connected to a sub-attr (those with R,G and B or XYZ at the end.)